Reimplemented architecture using Actor framework.
Currently this doesn't contain any of the features of the original
version but it should be easy enough to add them on top of this.

This architecture is a lot more maintainable and makes adding new
functionality much easier without having to muck around with lots of
closures, massive future types and lifetimes.