Re-implemented functionality with new architecture.
Everything works as before. This architecture makes continued
development easier - this is particularly useful for team members
that have less experience with Rust and the event bus project
where the old architecture would have been incredibly painful.

It is also more reliable, due to the actor architecture, it is far
easier to implement functionality and have all the edge cases handled in
a predicable way, in particular how failures in one component would
affect another component.

There are however less tests, this is due to the architecture being more
modular where messages are passed around between actors - this means
that each actor handles a smaller piece of functionality to the point
where most tests didn't make sense in the same way they did before. For
now, I've made tests to ensure the schemas are correct.