Improved redelivery of events.
Previously, when a client disconnects with unacknowledged events then
those events are added to the propagate event system to be resent.

However, if no service was available of that type then those events
would be lost. Further, there was no guarantee that a event that
was being retried would not be sent to a service of a client type that
did successfully process that event.

This commit modifies the system so that when a event is not acknowledged
and then a resend is attempted, if that resend fails, that event is
added to structure for that client type.

As the only case where an event can be added to this structure is when
there are no other events of that type for the event to be sent to, when
a new client connects and registers at that point we can check if
there are any events that need resent for that client type and then
attempt to send them.

Further, the initial resend attempt on unacknowledgement now ensures
that it is only resent to that client type.